Rework FFT system; Replace chfft with rustfft; enable fft_rustfft by default
Microbenchmarks (c.f. benches/benches.rs) run on two machines i have access to indicate
that rustfft is 10-50 times faster compared to chfft exspecially for large fft sizes. Thus i decided
to switch to rustfft even if its API is clunkier (which is reflected by the Cfft convenience
wrapper which most likely requires a lot more memory to fulfill the Fft interface ).
I'm considering exposing rustfft alongside the wrapper for non-rapid-prototyping
purposes where memory overhead is a concern but i'll keep it as it is for now.
The switch also means that from now on num-traits and num-complex are in step with
rustfft, because i consider it better practice to hold the dependencies at major/minor versions
when available